fix bug: multiple sessions on same file - only one session was updated on changes
fix bug: multiple sessions on same file - only one session was updated
on changes. if several browsers are open, and ask to tail the same file,
only a single session was updated on chages in this file. because the
cache-id was the file name, and the first thread who read the data,
deleted the entry from the cache.

changes: maintain client identifier as the key in the cache. client make
reads from cache using this id. he receives this id upon first call, on
startTailingFile().