chunks solution
chunks! client cannot read huge json, in case of big log file (>50M).
the JSON conversion threw OutOfMemory Error. so the data is divided into
chunks, and written to the cache in chunks. the ID of each element is
the "clientID" concatanated with the chunk-id. and there is a map that
manages all chunks per client.
size of a chunk is a param, read from a prop file (using Spring of
course, in CacheHolder.java)